Move flashCards flatMap to mongodb aggregate (#13)
This change makes the query for the flashcards much faster, because the
bottleneck in the current implementation is the flatMap operation we are
doing on the note's flashcards.

By moving this "flatMap" to the database, it can use whatever
optimizations they have in place to make it faster. On a deck with 2K
notes and 8K flashcards, it changed from ~1.3 seconds to ~30 milliseconds
to execute the query.