Updates scheduler algorithm to handle better repeat steps
This commit updates the scheduling algorithm to be more in
line with Anki's internal algorithm. The queues used in Anki
are not represented here because we have a different reviewing
model than it does (considering the whole day instead of an
interval of the next minutes), so some of those queues didn't
make sense and could be accomplished with only the flashcard status.

This also updates some of the references for the name "flashcard"
to be all in lowercase, which is preferred.